refactor: extract pixel_painter/furniture.rs + move anchors
- Move 5 furniture paint helpers (coffee_table, area_rug, side_table,
  pantry_table, pantry_chair) from mod.rs to furniture.rs
- Move character_anchor + compute_door_frame_idx to anchors.rs
- mod.rs stays as the orchestrator (~1308 lines, down from ~1500)
