+++ crates/ascii-agents/src/tui/pixel_painter/mod.rs
@@ -28,11 +28,13 @@ mod anchors;
 mod background;
 mod drawable;
 mod effects;
+mod furniture;
 mod palette;
 
+pub(in crate::tui) use anchors::character_anchor;
 use anchors::{
-    back_couch_anchor, seated_anchor, standing_at_desk_anchor, walking_anchor, walking_position,
-    waypoint_anchor, waypoint_rank_offset_x, with_breath,
+    back_couch_anchor, compute_door_frame_idx, seated_anchor, standing_at_desk_anchor,
+    walking_anchor, walking_position, waypoint_anchor, waypoint_rank_offset_x, with_breath,
 };
 use background::{
     dim_floor_overlay, paint_ceiling_pool, paint_clock, paint_corridor_runner,
@@ -83,257 +85,6 @@ pub(super) fn paint_character_at(
     blit_frame(cached, anchor.x, anchor.y, buf);
 }
 
-/// Low coffee table in front of the lounge couch. Wood top with darker
-/// trim along the front edge so it reads as a real piece of furniture,
-/// not just a brown rectangle.
-pub(super) fn paint_coffee_table(
-    buf: &mut RgbBuffer,
-    cx: u16,
-    cy: u16,
-    w: u16,
-    h: u16,
-    theme: &crate::tui::theme::Theme,
-) {
-    let top = theme.furniture.wood_top;
-    let trim = theme.furniture.wood_trim;
-    let min_x = cx.saturating_sub(w / 2);
-    let max_x = (cx + w / 2 + (w & 1)).min(buf.width);
-    let min_y = cy.saturating_sub(h / 2);
-    let max_y = (cy + h / 2 + (h & 1)).min(buf.height);
-    for y in min_y..max_y {
-        for x in min_x..max_x {
-            let on_front = y + 1 == max_y;
-            buf.put(x, y, if on_front { trim } else { top });
-        }
-    }
-}
-
-/// Meeting-room area rug — warm Persian-tone rectangle painted under
-/// the coffee table. Border ring in a darker shade so the rug reads as
-/// having a fringe/binding rather than a flat blob. Centred on `cx,cy`.
-pub(super) fn paint_area_rug(
-    buf: &mut RgbBuffer,
-    cx: u16,
-    cy: u16,
-    w: u16,
-    h: u16,
-    theme: &crate::tui::theme::Theme,
-) {
-    let rug_field = theme.furniture.rug_field;
-    let rug_trim = theme.furniture.rug_trim;
-    let rug_accent = theme.furniture.rug_accent;
-    let half_w = w as i32 / 2;
-    let half_h = h as i32 / 2;
-    for dy in 0..h as i32 {
-        for dx in 0..w as i32 {
-            let px = cx as i32 - half_w + dx;
-            let py = cy as i32 - half_h + dy;
-            if px < 0 || py < 0 || px >= buf.width as i32 || py >= buf.height as i32 {
-                continue;
-            }
-            let on_border = dx == 0 || dx == w as i32 - 1 || dy == 0 || dy == h as i32 - 1;
-            let on_inner_border = dx == 1 || dx == w as i32 - 2 || dy == 1 || dy == h as i32 - 2;
-            let color = if on_border {
-                rug_trim
-            } else if on_inner_border {
-                rug_accent
-            } else {
-                rug_field
-            };
-            buf.put(px as u16, py as u16, color);
-        }
-    }
-}
-
-/// Lounge side table — 7×4 wood block next to the viewing couch
-/// (opposite side from the floor lamp). Bumped from 5×3 to clear the
-/// skill's ~5-cell-wide subzone threshold. Carries a 3-cell magazine
-/// stack on top so the silhouette reads as "side table with a book".
-pub(super) fn paint_side_table(
-    buf: &mut RgbBuffer,
-    cx: u16,
-    cy: u16,
-    theme: &crate::tui::theme::Theme,
-) {
-    let top = theme.furniture.wood_top;
-    let trim = theme.furniture.wood_trim;
-    let mag = theme.furniture.magazine;
-    let mag_trim = theme.furniture.magazine_trim;
-    let w: i32 = 7;
-    let h: i32 = 4;
-    for dy in 0..h {
-        for dx in 0..w {
-            let px = cx as i32 - w / 2 + dx;
-            let py = cy as i32 - h / 2 + dy;
-            if px < 0 || py < 0 || px >= buf.width as i32 || py >= buf.height as i32 {
-                continue;
-            }
-            let on_bottom = dy == h - 1;
-            buf.put(px as u16, py as u16, if on_bottom { trim } else { top });
-        }
-    }
-    let mag_pixels: &[((i32, i32), Rgb)] = &[
-        ((-1, -1), mag),
-        ((0, -1), mag),
-        ((1, -1), mag),
-        ((-1, 0), mag_trim),
-        ((0, 0), mag_trim),
-        ((1, 0), mag_trim),
-    ];
-    for ((dx, dy), c) in mag_pixels {
-        let px = cx as i32 + dx;
-        let py = cy as i32 + dy;
-        if px >= 0 && py >= 0 && (px as u16) < buf.width && (py as u16) < buf.height {
-            buf.put(px as u16, py as u16, *c);
-        }
-    }
-}
-
-/// Pantry bistro table — round-ish wood top (rounded corners by skipping
-/// the 4 corner pixels) painted with the same warm wood palette as the
-/// coffee table so they read as the same furniture family.
-pub(super) fn paint_pantry_table(
-    buf: &mut RgbBuffer,
-    cx: u16,
-    cy: u16,
-    theme: &crate::tui::theme::Theme,
-) {
-    let top = theme.furniture.wood_top;
-    let trim = theme.furniture.wood_trim;
-    let w: i32 = 7;
-    let h: i32 = 4;
-    for dy in 0..h {
-        for dx in 0..w {
-            let on_corner = (dx == 0 || dx == w - 1) && (dy == 0 || dy == h - 1);
-            if on_corner {
-                continue;
-            }
-            let px = cx as i32 - w / 2 + dx;
-            let py = cy as i32 - h / 2 + dy;
-            if px < 0 || py < 0 || px >= buf.width as i32 || py >= buf.height as i32 {
-                continue;
-            }
-            let on_edge = dy == h - 1;
-            buf.put(px as u16, py as u16, if on_edge { trim } else { top });
-        }
-    }
-}
-
-pub(super) fn paint_pantry_chair(
-    buf: &mut RgbBuffer,
-    cx: u16,
-    cy: u16,
-    theme: &crate::tui::theme::Theme,
-) {
-    let seat = theme.furniture.chair_seat;
-    let trim = theme.furniture.chair_trim;
-    let put = |buf: &mut RgbBuffer, dx: i32, dy: i32, c: Rgb| {
-        let px = cx as i32 + dx;
-        let py = cy as i32 + dy;
-        if px >= 0 && py >= 0 && (px as u16) < buf.width && (py as u16) < buf.height {
-            buf.put(px as u16, py as u16, c);
-        }
-    };
-    put(buf, -1, -1, seat);
-    put(buf, 0, -1, seat);
-    put(buf, -1, 0, trim);
-    put(buf, 0, 0, trim);
-}
-
-/// How long the elevator's open/close transition takes. Used as both
-/// the opening ramp at the START of an agent's entry/exit window and
-/// the closing ramp at the END. 200 ms feels snappy without being
-/// abrupt — the half-open frame is visible for ~70 ms each way.
-const DOOR_TRANSITION_MS: u64 = 200;
-
-/// Compute the elevator door frame (0=closed, 1=half, 2=open) from
-/// the agents currently in flight. Stateless: each agent contributes
-/// a per-frame value based on how far through their entry/exit window
-/// they are; we take the MAX across all agents so the door is at
-/// least as open as the most-in-progress agent needs.
-fn compute_door_frame_idx(agents: &[AgentSlot], now: SystemTime) -> usize {
-    fn frame_for_progress(elapsed_ms: u64, total_ms: u64) -> usize {
-        // 0..200ms: opening (0 → 1 → 2)
-        if elapsed_ms < DOOR_TRANSITION_MS {
-            if elapsed_ms < DOOR_TRANSITION_MS / 2 {
-                1
-            } else {
-                2
-            }
-        } else if elapsed_ms + DOOR_TRANSITION_MS > total_ms {
-            // last 200ms: closing (2 → 1 → 0)
-            let remaining = total_ms.saturating_sub(elapsed_ms);
-            if remaining < DOOR_TRANSITION_MS / 2 {
-                0
-            } else {
-                1
-            }
-        } else {
-            // middle: fully open
-            2
-        }
-    }
-    let mut max_frame: usize = 0;
-    for a in agents {
-        if a.exiting_at.is_none() {
-            if let Ok(d) = now.duration_since(a.created_at) {
-                let ms = d.as_millis() as u64;
-                if ms < pose::ENTRY_ANIMATION_MS {
-                    max_frame = max_frame.max(frame_for_progress(ms, pose::ENTRY_ANIMATION_MS));
-                }
-            }
-        }
-        if let Some(exit_at) = a.exiting_at {
-            if let Ok(d) = now.duration_since(exit_at) {
-                let ms = d.as_millis() as u64;
-                // Use the same window the reducer uses to GC exiting
-                // slots so the door closes right as the agent's slot
-                // disappears.
-                let exit_window_ms =
-                    ascii_agents_core::state::reducer::EXIT_GRACE_WINDOW.as_millis() as u64;
-                if ms < exit_window_ms {
-                    max_frame = max_frame.max(frame_for_progress(ms, exit_window_ms));
-                }
-            }
-        }
-    }
-    max_frame
-}
-
-/// Current rendered position of an agent's character — derived from pose
-/// so labels can follow the character rather than staying anchored at the
-/// desk. Returns the top-left anchor of the character sprite. Uses
-/// `derive_with_routing` so labels track agents along their A* path
-/// instead of jumping the straight-line midpoint.
-#[allow(clippy::too_many_arguments)]
-pub(super) fn character_anchor(
-    agent: &AgentSlot,
-    layout: &Layout,
-    now: SystemTime,
-    router: &mut dyn Router,
-    overlay: &OccupancyOverlay,
-    history: &mut pose::PoseHistory,
-) -> Option<Point> {
-    use crate::tui::layout::WaypointKind;
-    let desk = *layout.home_desks.get(agent.desk_index)?;
-    let pose = pose::derive_with_routing(agent, now, layout, router, overlay, history)?;
-    let anchor = match pose {
-        Pose::SeatedIdle | Pose::SeatedThinking | Pose::SeatedTyping { .. } => seated_anchor(desk),
-        Pose::StandingAtDesk => standing_at_desk_anchor(desk),
-        Pose::AtWaypoint { wp, kind } => {
-            let wp_obj = layout.waypoints.get(wp)?;
-            match kind {
-                WaypointKind::Couch => back_couch_anchor(wp_obj.pos),
-                _ => waypoint_anchor(wp_obj.pos),
-            }
-        }
-        Pose::AimlessAt { dest } => waypoint_anchor(dest),
-        Pose::Walking {
-            from, to, t_x1000, ..
-        } => walking_anchor(walking_position(from, to, t_x1000)),
-    };
-    Some(anchor)
-}
 /// Pure pixel painting — no ratatui types, no terminal I/O. The signature
 /// is what any future non-terminal renderer (web canvas, PNG export, GIF
 /// capture) would call. Lives behind the `Renderer` trait in core if you
